Move command manager into commands folder
