+++ extensions/typescript-language-features/src/extension.browser.ts
@@ -13,7 +13,7 @@ import { noopLogDirectoryProvider } from './tsServer/logDirectoryProvider';
 import { ITypeScriptVersionProvider, TypeScriptVersion, TypeScriptVersionSource } from './tsServer/versionProvider';
 import { WorkerServerProcess } from './tsServer/serverProcess.browser';
 import API from './utils/api';
-import { CommandManager } from './utils/commandManager';
+import { CommandManager } from './commands/commandManager';
 import { TypeScriptServiceConfiguration } from './utils/configuration';
 import { PluginManager } from './utils/plugins';
